stable-sdk: consume execution cost from api
