+++ packages/stable-sdk/src/gasless/transfer.ts
@@ -6,7 +6,7 @@ import { ViemEvmClient } from "@stable-io/cctp-sdk-viem";
 import { Usdc, usdc, usdcContracts } from "@stable-io/cctp-sdk-definitions";
 
 import { Intent } from "../types/index.js";
-import { postTransferRequest } from "./api.js";
+import { postTransferRequest } from "../api/gasless.js";
 import { GaslessTransferData } from "src/methods/findRoutes/steps.js";
 
 export async function* transferWithGaslessRelay<
