+++ packages/stable-sdk/src/methods/findRoutes/routes/gasless.ts
@@ -11,7 +11,7 @@ import { SupportedEvmDomain, Corridor, CorridorStats } from "@stable-io/cctp-sdk
 import { ViemEvmClient } from "@stable-io/cctp-sdk-viem";
 
 import { RouteExecutionStep, gaslessTransferStep, signPermitStep } from "../steps.js";
-import { GetQuoteParams, getTransferQuote } from "../../../gasless/api.js";
+import { GetQuoteParams, getTransferQuote } from "../../../api/gasless.js";
 import { transferWithGaslessRelay } from "../../../gasless/transfer.js";
 import { calculateTotalCost, getCorridorFees } from "../fees.js";
 import { TransactionEmitter } from "../../../transactionEmitter.js";
